refactor: dedupe trade-slot helpers into src/util/tradeSlots.ts
normalizeSlotCount, toBoolArray, extractTradeItemIncluded were byte-identical copies in both DevServer.ts (proxy trade observer) and scripts/bridge/trade/index.ts (SDK trade bridge). Extract them plus DevServer's parseOfferSlots into a shared src/util/tradeSlots.ts that both import. Removes ~30 lines of duplication. DevServer.ts: 5,226 -> 5,168. Both copies verified live (observeTradePacket wired at the inspector subscription; installTrade called in the bridge). tsc clean.
